+++ src/cmem/ActivityControl/ActivityControlTypes.ts
@@ -1,28 +1,28 @@
 export interface IActivityStatus {
     // Optional project ID
-    project?: string
+    project?: string;
     // Optional task ID
-    task?: string
+    task?: string;
     // The activity name
-    activity: string
+    activity: string;
     // If the activity is currently running
-    isRunning: boolean
+    isRunning: boolean;
     // If the activity has failed
-    failed: boolean
+    failed: boolean;
     // The name of the status class
-    statusName: "Waiting" | "Finished" | "Idle" | "Running" | "Canceling"
+    statusName: "Waiting" | "Finished" | "Idle" | "Running" | "Canceling";
     // A number between 0 and 100
-    progress: number
+    progress: number;
     // More information corresponding to the status
-    message: string
+    message: string;
     // If the activity has been cancelled
-    cancelled: boolean
+    cancelled: boolean;
     // The concrete status ID
-    concreteStatus: "Cancelled" | "Failed" | "Successful" | "Not executed" | "Running" | "Waiting" | "Canceling"
+    concreteStatus: "Cancelled" | "Failed" | "Successful" | "Not executed" | "Running" | "Waiting" | "Canceling";
     // If there was an error, this contains the exception message
-    exceptionMessage?: string | null
+    exceptionMessage?: string | null;
     // The runtime in ms
-    runtime?: number
+    runtime?: number;
     // The start time as date time, e.g. "2021-09-07T09:34:53.153Z"
-    startTime?: string
+    startTime?: string;
 }