+++ src/components/AutocompleteField/AutoCompleteField.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useRef, useState} from "react";
 import {HTMLInputProps, IInputGroupProps, InputGroupProps, IPopoverProps, IRefObject} from "@blueprintjs/core";
 import {Suggest} from "@blueprintjs/select";
-import {Highlighter, IconButton, Menu, MenuItem, OverflowText, Spinner} from "../../../index";
+import {Highlighter, IconButton, Menu, MenuItem, OverflowText, Spinner} from "../../index";
 import {CLASSPREFIX as eccgui} from "../../configuration/constants";
 
 type SearchFunction<T extends any> = (value: string) => T[];
@@ -31,7 +31,7 @@ export interface IAutoCompleteFieldProps<T extends any, U extends any> {
      * @param value The value that has been converted with itemValueSelector.
      * @param e     The event
      */
-    onChange?(value: U, e?: React.SyntheticEvent<HTMLElement>);
+    onChange?(value: U, e?: React.SyntheticEvent<HTMLElement>): any;
 
     /**
      * The initial value for the auto-complete input field
@@ -198,7 +198,10 @@ export function AutoCompleteField<T extends any, U extends any>(props: IAutoComp
 
     useEffect(() => {
         setQueryToSelectedValue(selectedItem);
-    }, [selectedItem]);
+    },
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+        [selectedItem]
+    );
 
     useEffect(() => {
         if (!disabled && hasFocus) {
@@ -211,7 +214,10 @@ export function AutoCompleteField<T extends any, U extends any>(props: IAutoComp
                 setListLoading(false);
             };
         }
-    }, [hasFocus, query]);
+    },
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+        [hasFocus, query]
+    );
 
     const fieldWidthLimits = elementWidth(fieldRef);
 
@@ -233,13 +239,13 @@ export function AutoCompleteField<T extends any, U extends any>(props: IAutoComp
     }
 
     // Triggered when an item from the selection list gets selected
-    const onSelectionChange = (value, e) => {
+    const onSelectionChange = (value: any, e: any) => {
         setSelectedItem(value);
         onChange?.(itemValueSelector(value), e);
         setQueryToSelectedValue(value);
     };
 
-    const areEqualItems = (itemA, itemB) => itemValueSelector(itemA) === itemValueSelector(itemB);
+    const areEqualItems = (itemA: any, itemB: any) => itemValueSelector(itemA) === itemValueSelector(itemB);
 
     // Return the index of the item in the array based on the itemValueRenderer value
     const itemIndexOf = (arr: T[], searchItem: T): number => {
@@ -283,7 +289,7 @@ export function AutoCompleteField<T extends any, U extends any>(props: IAutoComp
     };
 
     // Renders the item in the selection list
-    const optionRenderer = (item, { handleClick, modifiers, query }) => {
+    const optionRenderer = (item: any, { handleClick, modifiers, query }: {handleClick: any, modifiers: any, query: any}) => {
         if (!modifiers.matchesPredicate) {
             return null;
         }