+++ src/components/Dialog/AlertDialog.tsx
@@ -3,19 +3,28 @@
 */
 
 import React from 'react';
-import {ClassNames as IntentClassNames} from "../../common/Intent";
-import SimpleDialog, { ISimpleDialogProps } from './SimpleDialog';
+import {Definitions as IntentStates, IntentTypes} from "../../common/Intent";
+import SimpleDialog, { SimpleDialogProps } from './SimpleDialog';
 
-export interface IAlertDialogProps extends ISimpleDialogProps {
-    // set to true if alert dialog displays a success message
-    success?: boolean;
-    // set to true if alert dialog displays a warning
-    warning?: boolean;
-    // set to true if alert dialog displays a strong message about errors or disruptive actions
-    danger?: boolean;
+export interface IAlertDialogProps extends Omit<SimpleDialogProps, "intent"> {
+  /**
+   * set to true if alert dialog displays a success message
+   */
+  success?: boolean;
+  /**
+   *  set to true if alert dialog displays a warning
+   */
+  warning?: boolean;
+  /**
+   * set to true if alert dialog displays a strong message about errors or disruptive actions
+   */
+  danger?: boolean;
 }
 
-
+/**
+ * Special element to display alert notification in modal dialogs.
+ * Inherits all properties from `SimpleDialog`, except `intent`.
+ */
 function AlertDialog({
     children,
     success=false,
@@ -23,16 +32,16 @@ function AlertDialog({
     danger=false,
     ...otherProps
 }: IAlertDialogProps) {
-    let intentLevel = IntentClassNames.INFO;
-    if (success) { intentLevel = IntentClassNames.SUCCESS; }
-    if (warning) { intentLevel = IntentClassNames.WARNING; }
-    if (danger) { intentLevel = IntentClassNames.DANGER; }
+    let intentLevel: IntentTypes = IntentStates.INFO;
+    if (success) { intentLevel = IntentStates.SUCCESS; }
+    if (warning) { intentLevel = IntentStates.WARNING; }
+    if (danger) { intentLevel = IntentStates.DANGER; }
 
     return (
         <SimpleDialog
             size="tiny"
             preventSimpleClosing={true}
-            intentClassName={intentLevel}
+            intent={intentLevel}
             {...otherProps}
         >
             {children}