+++ src/components/Menu/MenuItem.tsx
@@ -3,11 +3,12 @@ import {MenuItem as BlueprintMenuItem, MenuItemProps} from "@blueprintjs/core";
 import {CLASSPREFIX as eccgui} from "../../configuration/constants";
 import Icon from '../Icon/Icon';
 import { openInNewTab } from '../../common/utils/openInNewTab';
+import {ValidIconName} from "../Icon/canonicalIconNames";
 
 interface IProps {
     children?: React.ReactNode | React.ReactNode[]
     className?: string
-    icon?: string
+    icon?: ValidIconName | string[]
     // Props defined by the Blueprint component that should be forwarded
     internalProps?: Partial<MenuItemProps> & React.AnchorHTMLAttributes<HTMLAnchorElement>
     // FIXME: For backward compatibility, should be avoided in all code bases
@@ -24,13 +25,14 @@ function MenuItem({
                       href,
                       ...restProps
                   }: IProps) {
-    const actualHref = internalProps?.href ?? href
+    const actualHref = internalProps?.href ?? href;
+    const onClickHandler = internalProps?.onClick ?? onClick;
     return (
         <BlueprintMenuItem
             {...internalProps}
             {...restProps}
             href={actualHref}
-            onClick={(e) => openInNewTab(e, onClick, actualHref)}
+            onClick={(e) => openInNewTab(e, onClickHandler, actualHref)}
             className={`${eccgui}-menu__item ` + className}
             icon={
                 icon ? <Icon name={icon} /> : false