+++ src/components/Toolbar/ToolbarSection.tsx
@@ -1,6 +1,26 @@
 import React from "react";
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
+interface ToolbarSectionProps extends React.HTMLAttributes<HTMLDivElement> {
+    /**
+    * Is allowed to allocate more space than necessary to show all children.
+    */
+    canGrow?: boolean;
+    /**
+    * Allow element to get shrinked if there is not enough space to show all content.
+    * This only makes sense if it contains children that are elastic in its size.
+    */
+    canShrink?: boolean;
+    /**
+    * Hides content that overflows available space.
+    */
+    hideOverflow?: boolean;
+}
+
+/**
+ * Provides element to group toolbar elements together.
+ * It can be configured how dynamic it is displayed regarding provided space inside the `Toolbar` parent.
+ */
 function ToolbarSection({
     children,
     className = '',
@@ -8,7 +28,7 @@ function ToolbarSection({
     canShrink = false,
     hideOverflow = false,
     ...otherProps
-}: any) {
+}:ToolbarSectionProps) {
     return (
         <div
             {...otherProps}