+++ src/extensions/react-flow/nodes/NodeTools.tsx
@@ -2,7 +2,7 @@ import React, {memo, useState} from "react";import {
     IPopoverProps as IBlueprintPopoverProps,
     PopoverInteractionKind as BlueprintPopoverInteractionKind,
 } from "@blueprintjs/core";
-import {ContextOverlay, IconButton} from "../../../../index";
+import {ContextOverlay, IconButton} from "../../../index";
 import {CLASSPREFIX as eccgui} from "../../../configuration/constants";
 
 export interface NodeToolsProps extends IBlueprintPopoverProps {
@@ -34,7 +34,7 @@ export const NodeTools = memo(({
                     data-test-id={menuButtonDataTestId}
                     name={togglerElement}
                     text={togglerText}
-                    onmouseup={() => {
+                    onMouseUp={() => {
                         if (isOpened) { toggleIsOpened(false) };
                     }}/>
             ) : (